+++ ui/src/app/subnets/urls.ts
@@ -1,4 +1,4 @@
-import { generateLegacyURL } from "@maas-ui/maas-ui-shared";
+import { generateNewURL } from "@maas-ui/maas-ui-shared";
 
 import type { Fabric, FabricMeta } from "app/store/fabric/types";
 import type { Space, SpaceMeta } from "app/store/space/types";
@@ -23,13 +23,13 @@ const urls = {
 };
 
 const getFabricLink = (id?: Fabric["id"]): string | null =>
-  isId(id) ? generateLegacyURL(`/fabric/${id}`) : null;
+  isId(id) ? generateNewURL(`/fabric/${id}`) : null;
 const getSpaceLink = (id?: Space["id"]): string | null =>
-  isId(id) ? generateLegacyURL(`/space/${id}`) : null;
+  isId(id) ? generateNewURL(`/space/${id}`) : null;
 const getVLANLink = (id?: VLAN["id"]): string | null =>
-  isId(id) ? generateLegacyURL(`/vlan/${id}`) : null;
+  isId(id) ? generateNewURL(`/vlan/${id}`) : null;
 const getSubnetLink = (id?: Subnet["id"]): string | null =>
-  isId(id) ? generateLegacyURL(`/subnet/${id}`) : null;
+  isId(id) ? generateNewURL(`/subnet/${id}`) : null;
 
 export default urls;
 export { getFabricLink, getSpaceLink, getVLANLink, getSubnetLink };
