+++ ui/src/app/subnets/views/SubnetsList/SubnetsTable/FabricTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FabricTable";
